find_player関数の移動
